Added basic UI and Waves counter

+++ Assets/Scripts/HealthBar.cs
@@ -8,9 +8,9 @@ public class HealthBar : MonoBehaviour
     [SerializeField] private Image healthBarSprite;
 
     
-    public void UpdateHealthBar(float currentHealth,float maxHealth)
+    public void UpdateHealthBar(float currentHealth, float maxHealth)
     {
         Debug.Log("updaeHelath bar");
-        healthBarSprite.fillAmount = (float)currentHealth / (float)maxHealth;
+        healthBarSprite.fillAmount = currentHealth / maxHealth;
     }
 }
